Addressed review comments  - Introduced a new ParsedResponse and ParsedRequest struct
