+++ monoio-http/src/h2/frame/headers.rs
@@ -974,9 +974,6 @@ fn decoded_header_size(name: usize, value: usize) -> usize {
 
 #[cfg(test)]
 mod test {
-    use std::iter::FromIterator;
-
-    use http::HeaderValue;
 
     use super::*;
     use crate::h2::{
